moved auth store out to /web instead of /web/app
